refactor: extract unique id migration helpers
